refactor(admin): unify request lists into one shared RequestsTable (v0.22.3)
The dashboard preview and the per-key list each inlined their own
request table with only 10 columns, so they drifted from the canonical
/admin/requests view (16 columns) — missing Key, Task, Complexity,
Fallbacks, TPS and Error.

Extract the reference table into lib/components/RequestsTable.svelte and
reuse it on all three pages (~360 lines of duplication deleted). Per-page
knobs: detailHref (row link / Back target), showKey (dropped on the
key-scoped page), and two mutually-exclusive key-cell modes — onKeyFilter
(/requests in-page filter <button>) or keyHref (dashboard <a> linking to
the filtered list, since the dashboard has no in-page filter).
