+++ apps/admin/src/routes/requests/+page.svelte
@@ -3,11 +3,9 @@
   import { goto, invalidateAll } from '$app/navigation';
   import { base } from '$app/paths';
   import type { RequestListItem } from '$lib/api/requests.js';
-  import { attemptCodeLabel } from '$lib/format/attempt-codes.js';
   import RangeFilter from '$lib/components/RangeFilter.svelte';
   import RefreshControl from '$lib/components/RefreshControl.svelte';
-  import TokensCell from '$lib/components/TokensCell.svelte';
-  import { formatTimestamp, formatTps, formatUsd } from '$lib/format.js';
+  import RequestsTable from '$lib/components/RequestsTable.svelte';
   import { paginationItems } from '$lib/pagination.js';
   import {
     DEFAULT_FILTERS,
@@ -165,34 +163,6 @@
     const search = filtersToSearch(data.filters);
     const from = `${base}/requests${search ? `?${search}` : ''}`;
     return `${base}/requests/${traceId}?from=${encodeURIComponent(from)}`;
-  }
-
-  // Navigate when the row is clicked, EXCEPT when the click originates on an
-  // inner control (the request-id link handles its own navigation; the key cell
-  // is a filter button) — closest('a, button') lets those handle the click.
-  function onRowClick(event: MouseEvent, traceId: string): void {
-    if ((event.target as HTMLElement).closest('a, button')) return;
-    void goto(detailHref(traceId));
-  }
-
-  // Recorded time for the "timestamp" column: format the ISO ts in the viewer's
-  // local zone (shared helper); '—' when the record carried none (legacy row).
-  function formatTs(ts: string): string {
-    return formatTimestamp(ts) || '—';
-  }
-
-  // Distinct label styling per classification-stage decision layer (Principle 5).
-  function decidedByClass(d: RequestListItem['decided_by']): string {
-    switch (d) {
-      case 'rules':
-        return 'badge-rules';
-      case 'eval':
-        return 'badge-eval';
-      case 'fallback':
-        return 'badge-fallback';
-      default: // default
-        return 'badge-neutral';
-    }
   }
 </script>
 
@@ -358,17 +328,6 @@
     </span>
   </div>
 
-  <!-- Key cell contents (name + prefix subtitle, or bare prefix). Shared between
-       the clickable filter button and the non-clickable legacy fallback. -->
-  {#snippet keyLabel(r: RequestListItem)}
-    {#if r.key_name}
-      <span class="text-ink-strong" title={r.key_prefix}>{r.key_name}</span>
-      <code class="block font-mono text-xs text-ink-muted">{r.key_prefix}</code>
-    {:else}
-      <code class="font-mono text-ink-strong">{r.key_prefix}</code>
-    {/if}
-  {/snippet}
-
   {#if data.items.length === 0}
     <div data-testid="requests-empty" class="empty-state">
       <p class="font-medium text-ink-body">{$t('No requests match these filters.')}</p>
@@ -377,143 +336,7 @@
       </p>
     </div>
   {:else}
-    <div class="table-wrap">
-      <table class="table-base">
-        <thead class="table-head">
-          <tr>
-            <th class="px-3 py-2" title={$t('The unique trace ID recorded for this request.')}
-              >{$t('Request ID')}</th
-            >
-            <th class="px-3 py-2" title={$t('When the gateway received the request.')}
-              >{$t('Time')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t('The API key that authenticated the client, shown by prefix only.')}
-              >{$t('Key')}</th
-            >
-            <th class="px-3 py-2" title={$t('The model the client asked for in its request.')}
-              >{$t('Requested model')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t('Task type the classifier detected (e.g. coding, json, vision).')}
-              >{$t('Task')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t('Estimated request difficulty used to pick a quality tier.')}
-              >{$t('Complexity')}</th
-            >
-            <th class="px-3 py-2" title={$t('Which classification layer chose the lane.')}
-              >{$t('Decided by')}</th
-            >
-            <th class="px-3 py-2" title={$t('The quality/cost tier the request was routed to.')}
-              >{$t('Lane')}</th
-            >
-            <th class="px-3 py-2" title={$t('The model that actually handled the request.')}
-              >{$t('Served model')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t('How many fallback models were tried before one succeeded.')}
-              >{$t('Fallbacks')}</th
-            >
-            <th class="px-3 py-2" title={$t('Whether the request succeeded or returned an error.')}
-              >{$t('Status')}</th
-            >
-            <th class="px-3 py-2" title={$t('End-to-end response time in milliseconds.')}
-              >{$t('Latency')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t(
-                'Generation throughput: output tokens per second. Only measured for streamed responses.',
-              )}>{$t('TPS')}</th
-            >
-            <th
-              class="px-3 py-2"
-              title={$t('Token usage: input / output, with cached input tokens.')}
-              >{$t('Tokens')}</th
-            >
-            <th class="px-3 py-2" title={$t('Estimated cost of the request in US dollars.')}
-              >{$t('Cost')}</th
-            >
-            <th class="px-3 py-2" title={$t('The error class, if the request failed.')}
-              >{$t('Error')}</th
-            >
-          </tr>
-        </thead>
-        <tbody>
-          {#each data.items as r (r.trace_id)}
-            <!-- The whole row links to the detail page; the request-id cell keeps a
-                 real <a> for keyboard / open-in-new-tab. -->
-            <!-- svelte-ignore a11y_click_events_have_key_events a11y_no_static_element_interactions -->
-            <tr
-              data-testid="request-row"
-              class="table-row cursor-pointer"
-              onclick={(e) => onRowClick(e, r.trace_id)}
-            >
-              <td class="px-3 py-2">
-                <a
-                  class="link-inline block max-w-[7rem] truncate font-mono text-ink-strong lg:max-w-none"
-                  href={detailHref(r.trace_id)}
-                  title={r.trace_id}>{r.trace_id}</a
-                >
-              </td>
-              <td class="px-3 py-2 text-ink-body">{formatTs(r.ts)}</td>
-              <td class="px-3 py-2">
-                {#if r.key_id}
-                  <!-- Click the key to scope the list to it (sets the ?key_id
-                       filter). A <button>, so onRowClick lets it handle the click
-                       instead of opening the detail page. -->
-                  <button
-                    type="button"
-                    data-testid="key-filter"
-                    class="block text-left hover:underline"
-                    title={$t('Filter by this key')}
-                    onclick={() => go({ keyId: r.key_id })}
-                  >
-                    {@render keyLabel(r)}
-                  </button>
-                {:else}
-                  {@render keyLabel(r)}
-                {/if}
-              </td>
-              <td class="px-3 py-2 text-ink-body">{r.requested_model ?? '—'}</td>
-              <td class="px-3 py-2 text-ink-body">{r.task_type || '—'}</td>
-              <td class="px-3 py-2 text-ink-body">{r.complexity || '—'}</td>
-              <td class="px-3 py-2">
-                <span data-testid="decided-by" class={decidedByClass(r.decided_by)}
-                  >{r.decided_by}</span
-                >
-              </td>
-              <td class="px-3 py-2 text-ink-body">{r.lane || '—'}</td>
-              <td class="px-3 py-2 text-ink-body">{r.final_model ?? '—'}</td>
-              <td class="px-3 py-2 text-ink-body">{r.fallback_count}</td>
-              <td class="px-3 py-2">
-                {#if r.status === 'error'}
-                  <span class="badge-error">{$t('error')}</span>
-                {:else}
-                  <span class="badge-ok">{$t('ok')}</span>
-                {/if}
-              </td>
-              <td class="px-3 py-2 font-mono text-ink-body">{r.latency_ms}ms</td>
-              <td data-testid="cell-tps" class="px-3 py-2 font-mono text-ink-body"
-                >{formatTps(r.tps)}</td
-              >
-              <td class="px-3 py-2"><TokensCell usage={r.usage} /></td>
-              <td class="px-3 py-2 font-mono text-ink-body">{formatUsd(r.cost_usd)}</td>
-              <td
-                class="px-3 py-2 {r.error_class ? 'text-red-600' : 'text-ink-muted'}"
-                title={r.error_class ?? undefined}
-                >{r.error_class ? $t(attemptCodeLabel(r.error_class)) : '—'}</td
-              >
-            </tr>
-          {/each}
-        </tbody>
-      </table>
-    </div>
+    <RequestsTable items={data.items} {detailHref} onKeyFilter={(keyId) => go({ keyId })} />
 
     <!-- Pagination footer: rows-per-page on the left, numbered pages + Prev/Next on
          the right. `total` reflects the active filters so the counts stay consistent
