+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "helm-api",
-  "version": "0.22.2",
+  "version": "0.22.3",
   "private": true,
   "type": "module",
   "engines": {
